correction Sonar
