+++ src/main/analyses/temporelles/AnalyseTemporelle.component.js
@@ -88,17 +88,19 @@ export default class AnalyseTemporelle extends Component {
                     <Grid2 md={2} direction={"row-reverse"}>
                         {
                             this.state.listeCategories != null ?
-                                <AnalyseTemporelleFiltre listeCategories={this.state.listeCategories}
-                                                         onFilterChange={this.onFilterChange}/> : <CircularProgress/>
+                                <>
+                                    <AnalyseTemporelleFiltre listeCategories={this.state.listeCategories}
+                                                             onFilterChange={this.onFilterChange}/>
+                                    <FormControlLabel id="Soldes" key="Soldes" label="Soldes"
+                                                      control={<Checkbox id="Soldes" defaultChecked={false}
+                                                                         icon={<RadioButtonUnchecked/>}
+                                                                         checkedIcon={<CheckCircle/>}/>}
+                                                      style={{color: "#FFFFFF"}}
+                                                      onChange={this.onFilterSoldesChange}/>
+                                </>
+                                :
+                                <CircularProgress/>
                         }
-
-                        <FormControlLabel id="Soldes" key="Soldes"
-                                          control={<Checkbox id="Soldes" defaultChecked={false}
-                                                             icon={<RadioButtonUnchecked/>}
-                                                             checkedIcon={<CheckCircle/>}/>}
-                                          label="Soldes"
-                                          style={{color: "#FFFFFF"}}
-                                          onChange={this.onFilterSoldesChange}/>
                     </Grid2>
                 </Grid2>
                 <Divider variant="middle" sx={{margin: 1}}/>
